One giant step...
Added in a fake service which does a 'search'. It simualtes searching
by pausing for some random number of seconds then returning 'results'.

Using RxJS features like combineLatest & switchMap to make sure that
we search if either critiera changes and that we only ever care about
the 'last' one that we sent.